% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_function.R
\name{eb_number}
\alias{eb_number}
\title{eb_number}
\usage{
eb_number(LTABC, LTCTWO, abcweight, ctwodeweight)
}
\value{
eb_predictor output for UUP
}
\description{
This function loads the census data around class, dices by ward
calculates based on information provided from lucid talk for unionists
}
